Rename `DealParams` and `DealResponse` to convey relation to "proposal"
`DealParams` in Boost is the struct by which deal proposal configuration
is passed onto a SP. The response to it, either acceptance or rejection,
is then captured by a struct named `DealResponse`. The original names
could be clearer in terms of expressing that they relate to deal
proposal negotiation between the SP and a client.

Rename them to `DealProposal` and `DealProposalResponse` for better
human readability. The struct names has no impact on the wire protocol
and this should impose no breaking changes on Boost libp2p protocols.